requirements.txt
flask
flask_httpauth
flask-sqlalchemy